Make backwards compatible, and remove old code
