+++ Sources/IncitoDocument.swift
@@ -1,0 +1,39 @@
+//
+//  IncitoDocument.swift
+//  Incito-iOS
+//
+//  Created by Laurie Hufford on 08/11/2019.
+//  Copyright © 2019 ShopGun. All rights reserved.
+//
+
+import Foundation
+
+public struct IncitoEnvironment {
+    /// A list of all the incito schema versions supported by this library.
+    public static let supportedVersions: [String] = ["1.0.0"]
+}
+
+public struct IncitoDocument: Identifiable {
+
+    public struct Element {
+        public typealias Identifier = String
+        
+        public var id: Identifier
+        public var role: String?
+        public var meta: [String: JSONValue] = [:]
+        public var featureLabels: [String] = []
+        public var link: URL?
+        public var title: String?
+    }
+    
+    public typealias Identifier = String
+    
+    public var id: Identifier
+    public var version: String
+    public var backgroundColor: UIColor? = nil
+    public var meta: [String: JSONValue] = [:]
+    public var locale: String?
+    public var elements: [Element] = []
+    
+    public var json: String
+}
