--- Sources/UIKit/IncitoEnvironment+UIKit.swift
@@ -1,17 +1,0 @@
-//
-//  ┌────┬─┐         ┌─────┐
-//  │  ──┤ └─┬───┬───┤  ┌──┼─┬─┬───┐
-//  ├──  │ ╷ │ · │ · │  ╵  │ ╵ │ ╷ │
-//  └────┴─┴─┴───┤ ┌─┴─────┴───┴─┴─┘
-//               └─┘
-//
-//  Copyright (c) 2018 ShopGun. All rights reserved.
-
-import Foundation
-
-extension IncitoEnvironment {
-    public static var current = IncitoEnvironment(
-        imageLoader: IncitoDataStore.shared,
-        fontLoader: IncitoDataStore.shared
-    )
-}
