--- Sources/UIKit/UIView+RoundedCorners.swift
@@ -1,103 +1,0 @@
-//
-//  ┌────┬─┐         ┌─────┐
-//  │  ──┤ └─┬───┬───┤  ┌──┼─┬─┬───┐
-//  ├──  │ ╷ │ · │ · │  ╵  │ ╵ │ ╷ │
-//  └────┴─┴─┴───┤ ┌─┴─────┴───┴─┴─┘
-//               └─┘
-//
-//  Copyright (c) 2018 ShopGun. All rights reserved.
-
-import UIKit
-
-/// https://stackoverflow.com/a/53128198/318834
-
-extension UIView {
-    @discardableResult
-    public func roundCorners(topLeft: CGFloat = 0, topRight: CGFloat = 0, bottomLeft: CGFloat = 0, bottomRight: CGFloat = 0) -> CGPath {
-        let maskPath = UIBezierPath(
-            roundedRect: bounds,
-            topLeft: topLeft,
-            topRight: topRight,
-            bottomLeft: bottomLeft,
-            bottomRight: bottomRight
-        )
-        
-        let shape = CAShapeLayer()
-        shape.path = maskPath.cgPath
-        layer.mask = shape
-        return shape.path!
-    }
-}
-extension UIBezierPath {
-    
-    convenience init(roundedRect rect: CGRect, topLeft: CGFloat = 0, topRight: CGFloat = 0, bottomLeft: CGFloat = 0, bottomRight: CGFloat = 0) {
-        self.init(
-            roundedRect: rect,
-            topLeftRadius: CGSize(width: topLeft, height: topLeft),
-            topRightRadius: CGSize(width: topRight, height: topRight),
-            bottomLeftRadius: CGSize(width: bottomLeft, height: bottomLeft),
-            bottomRightRadius: CGSize(width: bottomRight, height: bottomRight)
-        )
-    }
-    
-    convenience init(roundedRect rect: CGRect, topLeftRadius: CGSize = .zero, topRightRadius: CGSize = .zero, bottomLeftRadius: CGSize = .zero, bottomRightRadius: CGSize = .zero) {
-        
-        self.init()
-        
-        let path = CGMutablePath()
-        
-        let topLeft = rect.origin
-        let topRight = CGPoint(x: rect.maxX, y: rect.minY)
-        let bottomRight = CGPoint(x: rect.maxX, y: rect.maxY)
-        let bottomLeft = CGPoint(x: rect.minX, y: rect.maxY)
-        
-        if topLeftRadius != .zero {
-            path.move(to: CGPoint(x: topLeft.x + topLeftRadius.width, y: topLeft.y))
-        } else {
-            path.move(to: topLeft)
-        }
-        
-        if topRightRadius != .zero {
-            path.addLine(to: CGPoint(x: topRight.x - topRightRadius.width, y: topRight.y))
-            path.addCurve(
-                to: CGPoint(x: topRight.x, y: topRight.y + topRightRadius.height),
-                control1: CGPoint(x: topRight.x, y: topRight.y),
-                control2: CGPoint(x: topRight.x, y: topRight.y + topRightRadius.height)
-            )
-        } else {
-            path.addLine(to: topRight)
-        }
-        
-        if bottomRightRadius != .zero{
-            path.addLine(to: CGPoint(x: bottomRight.x, y: bottomRight.y-bottomRightRadius.height))
-            path.addCurve(
-                to: CGPoint(x: bottomRight.x - bottomRightRadius.width, y: bottomRight.y),
-                control1: CGPoint(x: bottomRight.x, y: bottomRight.y),
-                control2: CGPoint(x: bottomRight.x - bottomRightRadius.width, y: bottomRight.y)
-            )
-        } else {
-            path.addLine(to: bottomRight)
-        }
-        
-        if bottomLeftRadius != .zero {
-            path.addLine(to: CGPoint(x: bottomLeft.x + bottomLeftRadius.width, y: bottomLeft.y))
-            path.addCurve(
-                to: CGPoint(x: bottomLeft.x, y: bottomLeft.y - bottomLeftRadius.height),
-                control1: CGPoint(x: bottomLeft.x, y: bottomLeft.y),
-                control2: CGPoint(x: bottomLeft.x, y: bottomLeft.y-bottomLeftRadius.height)
-            )
-        } else {
-            path.addLine(to: bottomLeft)
-        }
-        
-        if topLeftRadius != .zero{
-            path.addLine(to: CGPoint(x: topLeft.x, y: topLeft.y+topLeftRadius.height))
-            path.addCurve(to: CGPoint(x: topLeft.x+topLeftRadius.width, y: topLeft.y) , control1: CGPoint(x: topLeft.x, y: topLeft.y) , control2: CGPoint(x: topLeft.x+topLeftRadius.width, y: topLeft.y))
-        } else {
-            path.addLine(to: CGPoint(x: topLeft.x, y: topLeft.y))
-        }
-        
-        path.closeSubpath()
-        cgPath = path
-    }
-}
