+++ Sources/incito-webview.html
@@ -41,6 +41,15 @@
             return absoluteBoundsOfElement(document.querySelector("[data-id='"+elId+"']"));
         }
 
+        <!-- Returns array of (non-null) id strings for all the elements at the specified point -->
+        function getElementIdsAtPoint(x, y) {
+            return [...document.elementsFromPoint(x, y)]
+                .map(el => el.getAttribute('data-id'))
+                .filter(el => el != null)
+        }
+
+        <!-- Private Funcs -->
+        
         function absoluteBoundsOfElement(element) {
             var bounds = element.getBoundingClientRect();
 
@@ -59,12 +68,32 @@
             };
         }
 
-        <!-- Returns array of (non-null) id strings for all the elements at the specified point -->
-        <!-- NOTE: this only works in iOS 11 and above -->
-        function getElementIdsAtPoint(x, y) {
-            return [...document.elementsFromPoint(x, y)]
-                .map(el => el.getAttribute('data-id'))
-                .filter(el => el != null)
+        <!-- Polyfill for `elementsFromPoint` for older versions of ios safari -->
+        if (typeof document !== 'undefined' && typeof document.elementsFromPoint === 'undefined') {
+            document.elementsFromPoint = elementsFromPointPolyfill;
+        }
+        
+        function elementsFromPointPolyfill(x, y) {
+            var elements = [];
+            var pointerEvents =[];
+            var el;
+
+            do {
+                if (el !== document.elementFromPoint(x, y)) {
+                    el = document.elementFromPoint(x, y);
+                    elements.push(el);
+                    pointerEvents.push(el.style.pointerEvents);
+                    el.style.pointerEvents = 'none';
+                } else {
+                    el = null;
+                }
+            } while (el);
+
+            for (var i = 0; i < elements.length; i++) {
+                elements[i].style.pointerEvents = pointerEvents[i];
+            }
+
+            return elements;
         }
 
     </script>
